pistachio: pll: Fix 32bit integer overflows and pll rate calculation in integer mode
This commit fixes two issues :-
1. This commit fixes 32bit integer overflows throughout the pll driver
(i.e. wherever the result of integer multiplication may exceed the
range of u32).

One of the functions affected by this problem is .recalc_rate. It
returns incorrect rate for some pll settings (not for all though)
which in turn results in the incorrect rate setup of pll's child
clocks.

2. Bug: .recalc_rate callback for the fractional pll doesn't take operating
mode into account when calculating pll rate. This bug results in
the incorrect pll rates when pll is operating in integer mode.

This commit fixes this bug by checking the operating mode before
calculating pll rate.

+++ drivers/clk/pistachio/clk.h
@@ -109,16 +109,20 @@ struct pistachio_fixed_factor {
 		.parent		= _pname,			\
 	}
 
+/*
+ * in order to avoid u32 multiplication overflow, declare all
+ * members of this structure as u64
+ */
 struct pistachio_pll_rate_table {
-	unsigned long fref;
-	unsigned long fout;
-	unsigned long fout_min;
-	unsigned long fout_max;
-	unsigned int refdiv;
-	unsigned int fbdiv;
-	unsigned int postdiv1;
-	unsigned int postdiv2;
-	unsigned int frac;
+	unsigned long long fref;
+	unsigned long long fout;
+	unsigned long long fout_min;
+	unsigned long long fout_max;
+	unsigned long long refdiv;
+	unsigned long long fbdiv;
+	unsigned long long postdiv1;
+	unsigned long long postdiv2;
+	unsigned long long frac;
 };
 
 #define INT_PLL_RATES(_fref, _fout, _refdiv, _fbdiv,		\
